+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - improve stored file matching search
+- fix wrong result highlighting for lines that have \r, \n chars
 
 ### v0.0.4, Jul 9, 2014
 - bugfix for incorrect result highlighting